check the attached assembly file
There are N players participating in a marathon. Their numbers and time in which they 
completed the marathon are stored in the memory. It is required to rearrange them in 
ascending order to find the winner.
The inputs are the number of players and two tables. The first table contains the player 
number and the second one contains their recorded time.
The outputs are two tables. The first one contains the player number arranges according to 
their times and the second table shows these times.